requirements.txt
docx2pdf==0.1.8
lxml==4.9.3
pdf2image==1.16.3
Pillow==10.0.0
python-docx==0.8.11
reportlab==4.0.4
tk==0.1.0
tqdm==4.66.1
